move model extensions out of view
